+++ Rebus.Tests/Persistence/SqlServer/JsonFileSubscriptionStorageBasicSubscriptionOperations.cs
@@ -1,0 +1,10 @@
+﻿using NUnit.Framework;
+using Rebus.Tests.Contracts.Subscriptions;
+
+namespace Rebus.Tests.Persistence.SqlServer
+{
+    [TestFixture, Category(Categories.SqlServer)]
+    public class JsonFileSubscriptionStorageBasicSubscriptionOperations : BasicSubscriptionOperations<JsonFileSubscriptionStorageFactory>
+    {
+    }
+}