+++ Rebus/Config/OptionsConfigurer.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Runtime.CompilerServices;
@@ -8,6 +7,7 @@ using Rebus.Extensions;
 using Rebus.Injection;
 using Rebus.Logging;
 using Rebus.Pipeline;
+using Rebus.Timeouts;
 
 namespace Rebus.Config
 {
@@ -45,7 +45,29 @@ namespace Rebus.Config
         }
 
         /// <summary>
-        /// Registers the given factory function as a resolve of the given <see cref="TService"/> service
+        /// Configures Rebus to use another endpoint as the timeout manager
+        /// </summary>
+        public OptionsConfigurer UseExternalTimeoutManager(string timeoutManagerAddress)
+        {
+            if (string.IsNullOrWhiteSpace(timeoutManagerAddress))
+            {
+                throw new ArgumentException(string.Format("Cannot use '{0}' as an external timeout manager address!", timeoutManagerAddress), "timeoutManagerAddress");
+            }
+
+            if (!string.IsNullOrWhiteSpace(_options.ExternalTimeoutManagerAddressOrNull))
+            {
+                throw new InvalidOperationException(string.Format("Cannot set external timeout manager address to '{0}' because it has already been set to '{1}' - please set it only once!  (this operation COULD have been accepted, but it is probably an indication of an error in your configuration code that this value is configured twice, so we figured it was best to let you know)", 
+                    timeoutManagerAddress, _options.ExternalTimeoutManagerAddressOrNull));
+            }
+
+            _injectionist.Register<ITimeoutManager>(c => new ThrowingTimeoutManager());
+            _options.ExternalTimeoutManagerAddressOrNull = timeoutManagerAddress;
+            
+            return this;
+        }
+
+        /// <summary>
+        /// Registers the given factory function as a resolver of the given primary implementation of the <typeparamref name="TService"/> service
         /// </summary>
         public void Register<TService>(Func<IResolutionContext, TService> resolverMethod)
         {
@@ -53,11 +75,11 @@ namespace Rebus.Config
         }
 
         /// <summary>
-        /// Registers the given factory function as a resolve of the given <see cref="TService"/> service
+        /// Registers the given factory function as a resolve of the given decorator of the <typeparamref name="TService"/> service
         /// </summary>
-        public void Decorate<TService>(Func<IResolutionContext, TService> factoryMethod)
+        public void Decorate<TService>(Func<IResolutionContext, TService> resolverMethod)
         {
-            _injectionist.Register(factoryMethod, isDecorator: true);
+            _injectionist.Decorate(resolverMethod);
         }
 
         /// <summary>
@@ -73,7 +95,7 @@ namespace Rebus.Config
             {
                 if (serviceType != typeof (IPipeline)) return;
 
-                _injectionist.Register(c =>
+                _injectionist.Decorate(c =>
                 {
                     var pipeline = c.Get<IPipeline>();
 
@@ -94,7 +116,7 @@ Receive pipeline:
 
 
                     return pipeline;
-                }, isDecorator: true);
+                });
             };
 
             return this;