+++ Rebus/Pipeline/PipelineAbsolutePosition.cs
@@ -1,0 +1,9 @@
+﻿namespace Rebus.Pipeline
+{
+    public enum PipelineAbsolutePosition
+    {
+        Front, 
+ 
+        Back
+    }
+}