+++ Rebus/Serialization/JsonSerializer.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
@@ -14,7 +15,15 @@ namespace Rebus.Serialization
     /// </summary>
     internal class JsonSerializer : ISerializer
     {
-        const string JsonUtf8ContentType = "application/json;charset=utf-8";
+        /// <summary>
+        /// Proper content type when a message has been serialized with this serializer (or another compatible JSON serializer) and it uses the standard UTF8 encoding
+        /// </summary>
+        public const string JsonUtf8ContentType = "application/json;charset=utf-8";
+
+        /// <summary>
+        /// Contents type when the content is JSON
+        /// </summary>
+        public const string JsonContentType = "application/json";
 
         static readonly JsonSerializerSettings DefaultSettings = new JsonSerializerSettings
         {
@@ -24,25 +33,37 @@ namespace Rebus.Serialization
         static readonly Encoding DefaultEncoding = Encoding.UTF8;
 
         readonly JsonSerializerSettings _settings;
+        readonly Encoding _encoding;
 
         public JsonSerializer()
+            : this(DefaultSettings, DefaultEncoding)
         {
-            _settings = DefaultSettings;
+        }
+
+        internal JsonSerializer(Encoding encoding)
+            : this(DefaultSettings, encoding)
+        {
         }
 
         internal JsonSerializer(JsonSerializerSettings jsonSerializerSettings)
+            : this(jsonSerializerSettings, DefaultEncoding)
+        {
+        }
+
+        internal JsonSerializer(JsonSerializerSettings jsonSerializerSettings, Encoding encoding)
         {
             _settings = jsonSerializerSettings;
+            _encoding = encoding;
         }
 
         public async Task<TransportMessage> Serialize(Message message)
         {
             var jsonText = JsonConvert.SerializeObject(message.Body, _settings);
-            var bytes = DefaultEncoding.GetBytes(jsonText);
+            var bytes = _encoding.GetBytes(jsonText);
             var headers = message.Headers.Clone();
             var messageType = message.Body.GetType();
             headers[Headers.Type] = messageType.GetSimpleAssemblyQualifiedName();
-            headers[Headers.ContentType] = JsonUtf8ContentType;
+            headers[Headers.ContentType] = string.Format("{0};charset={1}", JsonContentType, _encoding.HeaderName);
             return new TransportMessage(headers, bytes);
         }
 
@@ -50,15 +71,63 @@ namespace Rebus.Serialization
         {
             var contentType = transportMessage.Headers.GetValue(Headers.ContentType);
 
-            if (contentType != JsonUtf8ContentType)
+            if (contentType == JsonUtf8ContentType)
             {
-                throw new FormatException(string.Format("Unknown content type: '{0}' - must be '{1}' for the JSON serialier to work", contentType, JsonUtf8ContentType));
+                return GetMessage(transportMessage, _encoding);
             }
 
-            var bodyString = DefaultEncoding.GetString(transportMessage.Body);
-            var bodyObject = JsonConvert.DeserializeObject(bodyString, _settings);
+            if (contentType.StartsWith(JsonContentType))
+            {
+                var encoding = GetEncoding(contentType);
+                return GetMessage(transportMessage, encoding);
+            }
+
+            throw new FormatException(string.Format("Unknown content type: '{0}' - must be '{1}' for the JSON serialier to work", contentType, JsonUtf8ContentType));
+        }
+
+        Encoding GetEncoding(string contentType)
+        {
+            var charset = contentType
+                .Split(';')
+                .Select(token => token.Split('='))
+                .Where(tokens => tokens.Length == 2)
+                .FirstOrDefault(tokens => tokens[0] == "charset");
+
+            if (charset == null)
+            {
+                return _encoding;
+            }
+
+            var encodingName = charset[1];
+
+            try
+            {
+                return Encoding.GetEncoding(encodingName);
+            }
+            catch (Exception exception)
+            {
+                throw new FormatException(string.Format("Could not turn charset '{0}' into proper encoding!", encodingName), exception);
+            }
+        }
+
+        Message GetMessage(TransportMessage transportMessage, Encoding bodyEncoding)
+        {
+            var bodyString = bodyEncoding.GetString(transportMessage.Body);
+            var bodyObject = Deserialize(bodyString);
             var headers = transportMessage.Headers.Clone();
             return new Message(headers, bodyObject);
+        }
+
+        object Deserialize(string bodyString)
+        {
+            try
+            {
+                return JsonConvert.DeserializeObject(bodyString, _settings);
+            }
+            catch (Exception exception)
+            {
+                throw new FormatException(string.Format("Could not deserialize JSON text: '{0}'", bodyString), exception);
+            }
         }
     }
 }